fix: rename abstract int to int16

+++ model/cpp_writer.go
@@ -67,7 +67,7 @@ func (cw *CppWriter) GetField(_ int, field Field) string {
 func (cw *CppWriter) GetTypeName(typeName string) string {
 	// https://en.cppreference.com/w/cpp/language/types
 	switch typeName {
-	case "int":
+	case "int16":
 		return "short int"
 	case "int32":
 		return "int"
